requirements.txt
playwright==1.49.1
python-dateutil==2.8.2
